pointerevents/pointerevent_pointerOver.html

<!doctype html>
<html>
    <head>
        <title>Pointer Event: Dispatch pointerover. </title>
        <meta content="text/html; charset=UTF-8" http-equiv="Content-Type"/>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/"/>
        <meta name="assert" content="When a pointing device is moved into the hit test boundaries of an element, the pointerover event must be dispatched. "/>
        <!-- http://w3c-test.org/resources/testharness.js -->
	    <script src="http://w3c-test.org/resources/testharness.js"></script>
		<!-- Additional helper script for common checks across event types -->
        <script type="text/javascript" src="pointerevent_support.js"></script>
		<script type="text/javascript">
		    var detected_pointertypes = {};
		    var test_pointerEvent = async_test("pointerover is dispatched");  // set up test harness;
		    // showPointerTypes is defined in pointerevent_support.js
		    // Requirements: the callbakc function will reference the test_pointerEvent object and
		    // will fail unless the async_test is created with the var name "test_pointerEvent".
		    add_completion_callback(showPointerTypes);

		    function run() {
		        var target0 = document.getElementById("target0");

		        on_event(target0, "pointerover", function (event) {
		            test_pointerEvent.step(function () {
		                check_PointerEvent(event);
		                assert_equals(event.type, "pointerover", "Pointer Event received");
		            });

		            detected_pointertypes[event.pointerType] = true;
		            test_pointerEvent.done(); // complete tes
		        });
		    }
		</script>
		<style>
			div {
				margin: 0em;
				padding: 2em;
			}

			#target0 {
				touch-action: none;
				background: yellow;
				border: 1px solid orange;
			}

			#complete-notice {
				border: 1px solid #0a0;
				display: none;
			}

			#pointertype-log {
				font-weight: bold;
			}
		</style>
	</head>
	<body onload="run()">
		<h1>Pointer Event: Dispatch pointerover.</h1>
		<h4>Test Description: 
				When a pointing device is moved into the hit test boundaries of an element, the pointerover event must be dispatched.  
		</h4>
		<br />
		<div id="target0">
			Use mouse, touch or pen to hover or contact this element..
		</div>
		<div id="complete-notice">
			<p>Test complete:  Scroll to Summary to view Pass/Fail Results.</p>
			<p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
			<p>Refresh the page to run the tests again with a different pointer type.</p>
		</div>
		<div id="log"></div>
	</body>
</html>
